Merge branch feature/member-session-view into main

+++ app/Ajax/App/Meeting/Settlement/Fee.php
@@ -84,6 +84,7 @@ class Fee extends CallableClass
         $pagination = $this->rq()->page()->paginate($pageNumber, $perPage, $billCount);
 
         $html = $this->view()->render('tontine.pages.meeting.settlement.page', [
+            'session' => $this->session,
             'charge' => $this->charge,
             'bills' => $bills,
             'pagination' => $pagination,
@@ -115,6 +116,12 @@ class Fee extends CallableClass
      */
     public function addSettlement(int $billId)
     {
+        if($this->session->closed)
+        {
+            $this->notify->warning(trans('meeting.warnings.session.closed'));
+            return $this->response;
+        }
+
         $this->settlementService->createSettlement($this->charge, $this->session, $billId);
         // $this->notify->success(trans('session.settlement.created'), trans('common.titles.success'));
 
@@ -128,6 +135,12 @@ class Fee extends CallableClass
      */
     public function delSettlement(int $billId)
     {
+        if($this->session->closed)
+        {
+            $this->notify->warning(trans('meeting.warnings.session.closed'));
+            return $this->response;
+        }
+
         $this->settlementService->deleteSettlement($this->charge, $this->session, $billId);
         // $this->notify->success(trans('session.settlement.deleted'), trans('common.titles.success'));
